// packages/slack-api/src/api/rawTypes.ts
export interface RawUserProfile {
  api_app_id?: string;
  avatar_hash?: string;
  bot_id?: string;
  display_name?: string;
  email?: string;
  fields?: Record<string, { alt?: string; value?: string } | undefined>;
  image_192?: string;
  image_48?: string;
  image_72?: string;
  phone?: string;
  pronouns?: string;
  real_name?: string;
  start_date?: string;
  status_emoji?: string;
  status_text?: string;
  team?: string;
  title?: string;
}

export interface RawUser {
  color?: string;
  id: string;
  is_admin?: boolean;
  is_bot?: boolean;
  is_owner?: boolean;
  is_primary_owner?: boolean;
  last_seen?: number;
  name?: string;
  presence?: string;
  profile?: RawUserProfile;
  real_name?: string;
  team_id?: string;
  tz?: string;
  tz_label?: string;
  tz_offset?: number;
}

export interface RawBot {
  app_id?: string;
  icons?: { image_36?: string; image_48?: string; image_72?: string };
  id: string;
  name?: string;
}